decouble dat-manager from choo
